Fixes iPhone 6 Plus Zooming Bug
It appears that somehow, the panGestureRecognizer on the scrollview is
interfering with every other gesture recognizer in some cases after
zooming. Its state is Possible at that point, so it shouldn’t be, but
disabling it after zooming and re-enabling it before zooming fixes the
bug while keeping the same behavior. Unfortunately this seems to be
another “I’m not quite sure why this fixes it” workaround, but at least
this fixes it.

Fixes #5