Remove pointless, crashing performance test
It’s unclear what this test is testing, since the method it calls could only operate on a protocol mock which is going to return `nil` objects anyway. And this test crashes on iOS 8 and 9, probably because of something related to the protocol mock which I would bother debugging if I thought this test were valuable. Instead, 🔥 it.