Include all files in the bundle, not just PNGs, in Cocoapods resource bundle
This change has no functional effect at this time. However, now that we have a proper bundle in the repo, it seems proper to direct CocoaPods to package _all_ files in that bundle into its (equivalent) resource bundle.

This eliminates room for future error: if a non-PNG file is added to the bundle, before this change, it would appear to work with Carthage but not with CocoaPods.