Update NYTPhotoDismissalInteractionController.m
Fixes issue, when you trying to dismiss NYTPhotoViewer (being landscape), which presented by portrait-only controller.
In that case panGestureRecognizer.state turns to "cancelled". We should treat that state the same way as "ended", because by the fact it ended.

Probably, there can be many other cases, which will be fix by this line of code.

Can be reproduced: 
iOS >= 8.0

I can provide video with "steps to reproduce", if you need this.